Created the Emergency Resources page
Just a rough draft of the emergency resources page for now. Added a new route in the application navigator for the homescreen. Integrated the call functionality from react-native-phone-call to allow users to click on list items to call the number. It's not working properly at the moment. It will fire immediately and then not allow you to try again.